+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModInnerPlatform.kt
@@ -1,0 +1,8 @@
+package site.siredvin.peripheralworks.xplat
+
+import site.siredvin.peripheralium.storages.energy.EnergyUnit
+import site.siredvin.peripheralium.xplat.BaseInnerPlatform
+
+interface ModInnerPlatform : BaseInnerPlatform {
+    val commonEnergy: EnergyUnit
+}