+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModPlatform.kt
@@ -1,18 +1,18 @@
 package site.siredvin.peripheralworks.xplat
 
-import site.siredvin.peripheralium.xplat.BaseInnerPlatform
+import site.siredvin.peripheralium.storages.energy.EnergyUnit
 import site.siredvin.peripheralium.xplat.BasePlatform
 import site.siredvin.peripheralium.xplat.ModInformationTracker
 
 object ModPlatform : BasePlatform {
-    private var _IMPL: BaseInnerPlatform? = null
+    private var _IMPL: ModInnerPlatform? = null
     private val _informationTracker = ModInformationTracker()
 
-    fun configure(impl: BaseInnerPlatform) {
+    fun configure(impl: ModInnerPlatform) {
         _IMPL = impl
     }
 
-    override val baseInnerPlatform: BaseInnerPlatform
+    override val baseInnerPlatform: ModInnerPlatform
         get() {
             if (_IMPL == null) {
                 throw IllegalStateException("You should configure upw ModPlatform first")
@@ -22,4 +22,7 @@ object ModPlatform : BasePlatform {
 
     override val modInformationTracker: ModInformationTracker
         get() = _informationTracker
+
+    val commonEnergy: EnergyUnit
+        get() = baseInnerPlatform.commonEnergy
 }